A thorough revamp of grimoire.ca's look and feel.
The new look is based, loosely, on print aesthetics, and uses Baskerville (and
Menlo, for code) throughout, with Google's Buenard as a fallback for Windows
and Linux users.

I've retained the basic visual flow, but reworked it to be both more visible
and more readable, and to have less-crazy layout rules (no more
borders-as-gaps).